+++ winnow-grammar-macro/src/codegen/mod.rs
@@ -425,6 +425,12 @@ impl<'a> Codegen<'a> {
                 ))
                 .map(|(_, c)| c)
             },
+            "hex_digit0" => quote_spanned! {span=>
+                (ws, ::winnow::ascii::hex_digit0).map(|(_, s)| AsRef::<str>::as_ref(&s).to_string())
+            },
+            "hex_digit1" => quote_spanned! {span=>
+                (ws, ::winnow::ascii::hex_digit1).map(|(_, s)| AsRef::<str>::as_ref(&s).to_string())
+            },
             _ => {
                 if args.is_empty() {
                     quote_spanned! {span=> #rule_name }